refactor : 날씨 관련 API 분리 & 파일 명 수정 & fetch함수 axios함수로 변경
1. 날씨 관련 API 분리
: src 폴더 아래 [api]라는 폴더 아래 "fetchWeatherData.ts" 파일 생성 후 openWeatherMap API 관련 코드 이동

2. 파일 명 수정
: 컴포넌트 파일 명 index.tsx로 통일

3. fetch함수 axios함수로 변경
: featch 함수는 응답에 type이 return 되지 않기 때문에 axios 함수로 변경
